ai-pdf-editor/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
pdf2image==1.16.3
pdfplumber==0.10.2
openai==0.28.0
google-generativeai==0.3.0
Pillow==10.0.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3